Добавил автоматический скроллинг карты для демонстрации

+++ QtFarterMike/level1screen.h
@@ -14,7 +14,7 @@ private:
 
     Level1SpriteSheet FSpriteSheet;
 
-    int FOffsetX;
+    float FOffsetX;
 
 public:
     Level1Screen();
